Add Conveyor/conveyorEffect support to the factory
Performace fixes to the Conveyor and Conveyor Effect scripts using raisedUpdateable so there is no loop checking if not needed.

These now also return true if it is a client of there is > 0 shaders added. No point loading otherwise as it is a waste and it also stops server loading as this is wrong if not protected ;-)